Adds lazy filtering.
This required that we change the traversal strategy protocol to allow it to produce a different number of objects than it consumes.

The traversal also applies its strategy slightly differently, ensuring that it does not stop enumerating when it produces no objects unless it has also consumed all the available objects, since for(in) interprets a 0 count return as being the end of the enumeration.