Explain -[RXTuple dealloc]'s need and mechanism.
This method is non-obvious both as to why it is necessary and why assigning nil to each element succeeds. Therefore, comment to explain it, with a link to the ARC docs for more information about the semantics of strong references at assignment time.

h/t @rustle for suggesting this comment.